move `ivy.cov` from linear alegbra to statistical submodule (#18109)
