+++ ivy/functional/backends/numpy/experimental/linear_algebra.py
@@ -173,33 +173,3 @@ def cond(
 
 
 cond.support_native_out = False
-
-
-def cov(
-    x1: np.ndarray,
-    x2: np.ndarray = None,
-    /,
-    *,
-    rowVar: bool = True,
-    bias: bool = False,
-    ddof: Optional[int] = None,
-    fweights: Optional[np.ndarray] = None,
-    aweights: Optional[np.ndarray] = None,
-    dtype: Optional[np.dtype] = None,
-) -> np.ndarray:
-    if fweights is not None:
-        fweights = fweights.astype(np.int64)
-
-    return np.cov(
-        m=x1,
-        y=x2,
-        rowvar=rowVar,
-        bias=bias,
-        ddof=ddof,
-        fweights=fweights,
-        aweights=aweights,
-        dtype=dtype,
-    )
-
-
-cov.support_native_out = False
