+++ ivy_tests/test_ivy/test_functional/test_experimental/test_core/test_linalg.py
@@ -571,140 +571,3 @@ def test_cond(
         x=x[0],
         p=x[1],
     )
-
-
-@st.composite
-def _get_dtype_value1_value2_cov(
-    draw,
-    available_dtypes,
-    min_num_dims,
-    max_num_dims,
-    min_dim_size,
-    max_dim_size,
-    abs_smallest_val=None,
-    min_value=None,
-    max_value=None,
-    allow_inf=False,
-    exclude_min=False,
-    exclude_max=False,
-    large_abs_safety_factor=4,
-    small_abs_safety_factor=4,
-    safety_factor_scale="log",
-):
-    shape = draw(
-        helpers.get_shape(
-            allow_none=False,
-            min_num_dims=min_num_dims,
-            max_num_dims=max_num_dims,
-            min_dim_size=min_dim_size,
-            max_dim_size=max_dim_size,
-        )
-    )
-
-    dtype = draw(st.sampled_from(draw(available_dtypes)))
-
-    values = []
-    for i in range(2):
-        values.append(
-            draw(
-                helpers.array_values(
-                    dtype=dtype,
-                    shape=shape,
-                    abs_smallest_val=abs_smallest_val,
-                    min_value=min_value,
-                    max_value=max_value,
-                    allow_inf=allow_inf,
-                    exclude_min=exclude_min,
-                    exclude_max=exclude_max,
-                    large_abs_safety_factor=large_abs_safety_factor,
-                    small_abs_safety_factor=small_abs_safety_factor,
-                    safety_factor_scale=safety_factor_scale,
-                )
-            )
-        )
-
-    value1, value2 = values[0], values[1]
-
-    # modifiers: rowVar, bias, ddof
-    rowVar = draw(st.booleans())
-    bias = draw(st.booleans())
-    ddof = draw(helpers.ints(min_value=0, max_value=1))
-
-    numVals = None
-    if rowVar is False:
-        numVals = -1 if numVals == 0 else 0
-    else:
-        numVals = 0 if len(shape) == 1 else -1
-
-    fweights = draw(
-        helpers.array_values(
-            dtype="int64",
-            shape=shape[numVals],
-            abs_smallest_val=1,
-            min_value=1,
-            max_value=10,
-            allow_inf=False,
-        )
-    )
-
-    aweights = draw(
-        helpers.array_values(
-            dtype="float64",
-            shape=shape[numVals],
-            abs_smallest_val=1,
-            min_value=1,
-            max_value=10,
-            allow_inf=False,
-            small_abs_safety_factor=1,
-        )
-    )
-
-    return [dtype], value1, value2, rowVar, bias, ddof, fweights, aweights
-
-
-# cov
-@handle_test(
-    fn_tree="functional.ivy.experimental.cov",
-    dtype_x1_x2_cov=_get_dtype_value1_value2_cov(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_num_dims=1,
-        max_num_dims=2,
-        min_dim_size=2,
-        max_dim_size=5,
-        min_value=1,
-        max_value=1e10,
-        abs_smallest_val=0.01,
-        large_abs_safety_factor=2,
-        safety_factor_scale="log",
-    ),
-    test_gradients=st.just(False),
-    test_with_out=st.just(False),
-)
-def test_cov(
-    *,
-    dtype_x1_x2_cov,
-    test_flags,
-    backend_fw,
-    fn_name,
-    on_device,
-    ground_truth_backend,
-):
-    dtype, x1, x2, rowVar, bias, ddof, fweights, aweights = dtype_x1_x2_cov
-    helpers.test_function(
-        ground_truth_backend=ground_truth_backend,
-        input_dtypes=[dtype[0], dtype[0], "int64", "float64"],
-        test_flags=test_flags,
-        fw=backend_fw,
-        fn_name=fn_name,
-        on_device=on_device,
-        x1=x1,
-        x2=x2,
-        rowVar=rowVar,
-        bias=bias,
-        ddof=ddof,
-        fweights=fweights,
-        aweights=aweights,
-        return_flat_np_arrays=True,
-        rtol_=1e-2,
-        atol_=1e-2,
-    )
